backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
requests
bcrypt==4.0.1
